Space Invaders
This is a Space Invaders game. The user is representing the spaceship
and has to destroy all the aliens found on levels one to five. When
these aliens are defeated a boss monster will be his next opponent which
is found on level 6. A health power up and a speed boost icon will
appear on the screen to help the player throughout the entire game.